crowbar: Stop rebuilding node run_list depending on the node state
Whether a role should be handled as part of a chef run will now be
decided in the role recipes.

The benefit of this is that the roles will always be assigned to the
node, even if the roles are not used due to the state. This fixes
several issues, including the fact that the search results for nodes are
not reliable when a node is in some state that is not "readying",
"applying" or "ready".