crowbar: Rework automatic addition of dns/logging/ntp roles to nodes
We used to do this when the node was discovered only. We now do it when
the node is entering installed or readying state:

 - there's no need to add the client roles before the node is installed,
   as the discovery image takes care of the setup during discovery
 - by doing it in readying, we also make sure that nodes get the roles
   back in case they lose it for some reason

We also have an additional check that the node is allocated; this is
technically not required, but this makes the transition method more
generic (as the check for the state is actually done through the data
bag): non-allocated nodes will never get these roles, which is fine
since they don't need them.