ipmi: Merge ipmi-discover and ipmi-configure roles into ipmi role
Instead of having two roles, we just use the role recipe to do the right
thing depending on the state of the node. This is less confusing.

Also, add the role if the node enters discovering/readying. See 4444959f
for the rationale. The only difference is that we use discovering instead of
installed, because we need to discover the IPMI-related attributes
during discovery.