chef: Split uefi code in its own recipe and include it where needed
We move the crowbar-hacks inclusion from deployer-client to
provisioner-base, but we still include the UEFI-related bits (now split
in a recipe) in the case of the discovery image in deployer-client.

Rationale for moving this is that deployer-client will always run in all
cases, even when we only run a subset of roles, so we should minimize
the amount of work it does in general.