skip backup tests when ActionController::RoutingError occurs
this is a workaround for a bug in the Rails routing engine that
only happens sometimes and only when unit testing a custom controller.

I think it could have to do with the namespacing of crowbar in our
api/backups_controller.

There is some dynamic route/controller calculation in ActionDispatch's
routing see:
https://github.com/rails/rails/blob/v4.2.5/actionpack/lib/action_dispatch/routing/route_set.rb:62

most of the times the controller is calculated correctly when you read
the params at that point:
$ params
{:controller=>"api/backups", :action=>"upload"}

and only sometimes it looks like:
params
{:controller=>"api/crowbar/backups", :action=>"upload"}

which is creating an ActionDispatch::RoutingError:
  uninitialized constant Api::Crowbar::BackupsController

there is an issue on GitHub which is likely related, and the easiest fix
for now is to just skip the test if the error occurs.

a real fix to test the post route for uploading the backup would likely
require extensive stubbing within:
  ActionDispatch::Routing::RouteSet::Dispatcher