add api documentation for crowbar-init's upgrade related api
as crowbar-init doesn't support apipie there is now now way to nicely
document the endpoints that we have there.

and to not have the documentation in yet another place it might be
better to include it directly in the crowbar api docs.

the downside of that is that the methods don't live in the crowbar
codebase, and therefore need to be faked by adding an empty dummy
method